web/requirements.txt
gunicorn==20.1.0
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connection==2021.7.20
connexion==2.14.2
Flask==2.2.5
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pyparsing==3.0.9
pyrsistent==0.19.3
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.15
swagger-ui-bundle==0.0.9
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.2.3
psycopg2-binary==2.9.6
WTForms==3.1.1
